+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 # Changelog
 
+## 4.14.2 (2026-09-02)
+
+- [Memory] Augment recall results to return the actual facts in structured format
+
 ## 4.14.1 (2026-08-20)
 
-- Add list_users api to memory
+- [Memory] Add list_users api to memory
+- [Memory] Rename users to sessions
+- [Memory] Rename topic to resource
 
 ## 4.14.0 (2026-07-31)